Move groupers into their own directory
